rename tera-fullscreen-modal => tera-drilldown
